# docs/swagger.yaml
basePath: /api
definitions:
  controllers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.AuditLog:
    properties:
      action:
        type: string
      created_at:
        type: string
      entity_id:
        type: integer
      entity_type:
        type: string
      id:
        type: integer
      new_value:
        items:
          type: integer
        type: array
      old_value:
        items:
          type: integer
        type: array
      user_id:
        $ref: '#/definitions/models.NullInt64'
      user_name:
        type: string
    type: object
  models.AuditLogCreate:
    properties:
      action:
        type: string
      entity_id:
        type: integer
      entity_type:
        type: string
      new_value:
        items:
          type: integer
        type: array
      old_value:
        items:
          type: integer
        type: array
      user_id:
        $ref: '#/definitions/sql.NullInt64'
    required:
    - action
    - entity_id
    - entity_type
    type: object
  models.Client:
    properties:
      address:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      notes:
        type: string
      phone:
        type: string
      updated_at:
        type: string
      vat_number:
        type: string
    type: object
  models.ClientCreate:
    properties:
      address:
        type: string
      email:
        type: string
      name:
        type: string
      notes:
        type: string
      phone:
        type: string
      vat_number:
        type: string
    required:
    - name
    type: object
...
  models.ClientSearchResult:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.ClientUpdate:
    properties:
      address:
        type: string
      email:
        type: string
      name:
        type: string
      notes:
        type: string
      phone:
        type: string
      vat_number:
        type: string
    required:
    - name
    type: object
...
  models.NullInt64:
    properties:
      int64:
        format: int64
        type: integer
      valid:
        description: Valid is true if Int64 is not NULL
        type: boolean
    type: object
...
  sql.NullInt64:
    properties:
      int64:
        format: int64
        type: integer
      valid:
        description: Valid is true if Int64 is not NULL
        type: boolean
    type: object
host: localhost:5000
info:
  contact:
    email: support@aura-erp.com
    name: API Support
    url: http://www.aura-erp.com/support
  description: This is a powerful ERP backend for Aura.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Aura ERP API
  version: "1.0"
paths:
  /audit-log:
    get:
      consumes:
      - application/json
      description: Get a paginated list of system audit logs
      parameters:
      - default: 50
        description: Limit results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset results
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AuditLog'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List audit logs
      tags:
      - audit
    post:
      consumes:
      - application/json
      description: Manually create an audit log record
      parameters:
      - description: Audit log details
        in: body
        name: auditLog
        required: true
        schema:
          $ref: '#/definitions/models.AuditLogCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AuditLog'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create an audit log entry
      tags:
      - audit
  /audit-log/{entityType}/{entityId}:
    get:
      consumes:
      - application/json
      description: Get all audit logs associated with a specific entity type and ID
      parameters:
      - description: Entity Type
        in: path
        name: entityType
        required: true
        type: string
      - description: Entity ID
        in: path
        name: entityId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AuditLog'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get audit logs for an entity
      tags:
      - audit
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return a JWT token
      parameters:
      - description: Login credentials
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - auth
...
  /auth/verify:
    get:
      consumes:
      - application/json
      description: Verify the JWT token and return user info
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify token
      tags:
      - auth
  /clients:
    get:
      consumes:
      - application/json
      description: Get all clients in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Client'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List clients
      tags:
      - clients
    post:
      consumes:
      - application/json
      description: Create a new client
      parameters:
      - description: Client details
        in: body
        name: client
        required: true
        schema:
          $ref: '#/definitions/models.ClientCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a client
      tags:
      - clients
  /clients/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a client from the system
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a client
      tags:
      - clients
    get:
      consumes:
      - application/json
      description: Get a single client by its ID
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Client'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get client by ID
      tags:
      - clients
    put:
      consumes:
      - application/json
      description: Update an existing client's details
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated client details
        in: body
        name: client
        required: true
        schema:
          $ref: '#/definitions/models.ClientUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a client
      tags:
      - clients
  /clients/search:
    get:
      consumes:
      - application/json
      description: Search clients by name
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - default: 25
        description: Limit results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ClientSearchResult'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search clients
      tags:
      - clients
...
swagger: "2.0"